requirements.txt
flask==3.0.0
openpyxl==3.1.2
google-auth==2.23.3
google-api-python-client==2.108.0
pypdn==1.0.6
numpy==1.26.0
Pillow==10.0.1
gunicorn